Refactor to use MissingTransactionInfo for rosetta APIs
